Move API classes to its own group
