+++ CatsFoundation/CatsFoundation/Breeds Feature/Breed.swift
@@ -1,0 +1,10 @@
+//
+//  Copyright © 2021 Jesús Alfredo Hernández Alarcón. All rights reserved.
+//
+
+import Foundation
+
+struct Breed: Decodable, Equatable {
+    let id: String
+    let name: String
+}
