--- CatsFoundation/CatsFoundation/CatsFoundation.h
@@ -1,18 +1,0 @@
-//
-//  CatsFoundation.h
-//  CatsFoundation
-//
-//  Created by Jesús Alfredo Hernández Alarcón on 05/04/21.
-//
-
-#import <Foundation/Foundation.h>
-
-//! Project version number for CatsFoundation.
-FOUNDATION_EXPORT double CatsFoundationVersionNumber;
-
-//! Project version string for CatsFoundation.
-FOUNDATION_EXPORT const unsigned char CatsFoundationVersionString[];
-
-// In this header, you should import all the public headers of your framework using statements like #import <CatsFoundation/PublicHeader.h>
-
-
